cmd/refresh: implement new contract notifications
InsertContract in db/crud had to be massaged a bit to return whether an
inserted contract is new (which means another query because it's SQLite we're
talking about).

New contract notifications is not implemented in cmd/events because calling
getPeriodicals is an implementation detail and the side effect of also
refreshing contracts is not guaranteed.