misc: implement contract progress bar
Note that a rewards field has been added to the SoloContract protobuf message,
which is necessary for progress bar rendering for solo contracts. However,
historical SoloContract messages have not been extended to include this field;
it is possible to do so, but requires too much work for too little
gain. Therefore, time traveling to before this commit went into effect would
not display progress bars for solo contracts, even though all the necessary
data is present in the database.