db: migrate: add contract.expiry_month
At the current rate of leggacy contracts it is a good chance we will eventually
get two runs of the same contract in the same calendar year. Adding
expiry_month to future-proof such conflicts.

While we're at it, recalculate expiry_year in UTC, so that the database isn't
subtly wrong if moved to a different timezone.