db/crud: GetCoopMemberActivityStats: vastly improve performance...
in the presence of members who haven't been seen for a long time.

Turns out db queries are very expensive here, and a page size of 60 (2 hours at
a time at a rate of 2 min per refresh) can lead to hundreds of queries when
a contract has been in abandonment for, say, 2 weeks. At ~60ms per query it
could take 10s or more to finish scanning.

We now increase the page size to 1080 for 36 hours, and stop after the first
two pages because there's really not much point in knowing whether someone has
been offline for 72 hours or 720 hours.